验证通过

+++ logdel/main.go
@@ -121,26 +121,25 @@ func parseConf(confFile string, confData *Configurations) []Items {
 	return instances
 }
 
-func getLogFiles(logItems []Items) []string {
-	logFiles := make([]string, 0)
-
+func delLogFiles(logItems []Items, optDryRun bool) {
 	for _, item := range logItems {
-		// 根据配置项目中的date_formats和suffixes进行组合拼一个正则。
-		tmpRegexpString := ""
-
 		// 先是根据文件后缀名进行分类。
 		for _, logSuffix := range item.Suffixes {
+			// 创建临时数组，用于存放匹配到的文件。
+			logFiles := make([]string, 0)
+
 			for _, dateFormat := range item.DateFormats {
+				// 根据配置项目中的date_formats和suffixes进行组合拼一个正则。
+				regexpString := ""
+				tmpRegexpString := ""
+
 				tmpRegexpString += strings.ToLower(dateFormat) + strings.ToLower(logSuffix) + "$|"
 				tmpRegexpString += strings.ToLower(logSuffix) + strings.ToLower(dateFormat) + "$|"
 
-				tmpLogFiles := make([]string, 0)
-				regexpString := tmpRegexpString[:len(tmpRegexpString)-1]
+				regexpString = tmpRegexpString[:len(tmpRegexpString)-1]
 				regexpString = strings.ReplaceAll(regexpString, "yyyy", "\\d{4}")
 				regexpString = strings.ReplaceAll(regexpString, "mm", "\\d{2}")
 				regexpString = strings.ReplaceAll(regexpString, "dd", "\\d{2}")
-
-				// log.Infof("正则表达式：%s\n", regexpString)
 
 				for _, path := range item.Paths {
 					files, err := ioutil.ReadDir(path)
@@ -151,34 +150,50 @@ func getLogFiles(logItems []Items) []string {
 						log.Infof("读取\"%s\"目录成功。\n", path)
 					}
 
+					log.Infof("当前匹配模式：%#v\n", regexpString)
 					for _, file := range files {
 						if !file.IsDir() {
 							fileName := file.Name()
-
-							validString := regexp.MustCompile(regexpString)
-							if validString.MatchString(fileName) {
-								log.Infof("匹配成功：%s\n", fileName)
-								tmpLogFiles = append(tmpLogFiles, path+"/"+fileName)
-							} else {
-								log.Warnf("匹配失败：%s\n", fileName)
+							if strings.HasSuffix(strings.ToLower(fileName), logSuffix) {
+								validString := regexp.MustCompile(regexpString)
+								if validString.MatchString(fileName) {
+									// 匹配成功的文件则存入在临时的数组中。
+									log.Infof("匹配成功：%s\n", fileName)
+									logFiles = append(logFiles, path+"/"+fileName)
+								} else {
+									log.Warnf("匹配失败：%s\n", fileName)
+								}
 							}
 						}
 					}
+
+					// 只有匹配到的文件大于需要保留的文件数时，才执行。
+					if len(logFiles) > item.RemainedNum {
+						logFiles = logFiles[:len(logFiles)-item.RemainedNum+1]
+
+						if optDryRun {
+							for _, logFile := range logFiles {
+								log.Infof("当前运行在dry-run模式，仅显示被删除日志文件名：%s", logFile)
+							}
+						} else {
+							for _, logFile := range logFiles {
+								log.Infof("当前运行在删除模式，即将删除此日志文件：%s", logFile)
+								if err := os.Remove(logFile); err != nil {
+									log.Infof("文件删除失败：%s，原因：%#v\n", logFile, err)
+								} else {
+									log.Infof("文件删除成功：%s\n", logFile)
+								}
+							}
+						}
+					} else {
+						log.Warnf("根据当前规划匹配不到文件或匹配到的文件数量少于预设值。\n")
+					}
+					logFiles = make([]string, 0)
 				}
-				tmpLogFiles = tmpLogFiles[:len(tmpLogFiles)-item.RemainedNum+1]
-				logFiles = append(logFiles, tmpLogFiles...)
 			}
 		}
 	}
-	return logFiles
-}
 
-func delLogFiles(logFile string) {
-	if err := os.Remove(logFile); err != nil {
-		log.Infof("文件删除失败：%s，原因：%#v\n", logFile, err)
-	} else {
-		log.Infof("文件删除成功：%s\n", logFile)
-	}
 }
 
 func main() {
@@ -213,17 +228,7 @@ func main() {
 			} else {
 				for _, logItem := range logItems {
 					log.Infof("配置文件\"%s\"中items配置项：%#v。\n", confFile, logItem)
-					logFiles := getLogFiles(logItems)
-					if optDryRun {
-						for _, logFile := range logFiles {
-							log.Infof("当前运行在dry-run模式，仅显示被删除日志文件名：%s", logFile)
-						}
-					} else {
-						for _, logFile := range logFiles {
-							log.Infof("当前运行在删除模式，即将删除此日志文件：%s", logFile)
-							delLogFiles(logFile)
-						}
-					}
+					delLogFiles(logItems, optDryRun)
 				}
 			}
 		}
